notification-service/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.1.0
bcrypt==4.2.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
coverage==7.6.10
cryptography==44.0.1
deprecation==2.1.0
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.1.5
django-allauth==65.4.1
django-cors-headers==4.6.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
ecdsa==0.19.1
et-xmlfile==1.1.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gotrue==2.9.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
huepy==1.2.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
livereload==2.7.1
lizard==1.17.20
Mako==1.3.8
mando==0.7.1
markdown2==2.5.3
MarkupSafe==2.1.3
numpy==2.2.5
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
pluggy==1.5.0
postgrest==0.15.1
psycopg2==2.9.10
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
PySocks==1.7.1
pytest==8.3.5
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
radon==6.0.1
realtime==1.0.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.29.0
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
sqlparse==0.5.3
storage3==0.7.7
StrEnum==0.4.15
supabase==2.3.4
supafunc==0.3.1
tornado==6.4.2
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vite==1.5.2
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.1.3
whitenoise==6.9.0
wsproto==1.2.0
WTForms==3.2.1
